requirements.txt
backoff==2.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coinbase-advanced-py==1.2.1
cryptography==42.0.5
idna==3.6
numpy==1.26.4
pandas==2.2.1
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
uuid==1.30
websockets==12.0
